Moved classes from o.e.s.model.item to o.e.s.core
