+++ bundles/core/org.eclipse.smarthome.core.autoupdate/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Import-Package: org.apache.commons.lang,
  org.eclipse.smarthome.core.events,
  org.eclipse.smarthome.core.items,
  org.eclipse.smarthome.core.types,
- org.eclipse.smarthome.model.item.binding,
  org.osgi.framework,
  org.osgi.service.cm,
  org.osgi.service.component,
