+++ bundles/core/org.eclipse.smarthome.core.transform/META-INF/MANIFEST.MF
@@ -22,7 +22,6 @@ Import-Package: com.google.common.base,
  org.eclipse.smarthome.core.scriptengine.action,
  org.eclipse.smarthome.core.types,
  org.eclipse.smarthome.io.net.exec,
- org.eclipse.smarthome.model.item.binding,
  org.osgi.framework,
  org.osgi.service.cm,
  org.osgi.service.component,
