+++ bundles/core/org.eclipse.smarthome.core/src/main/java/org/eclipse/smarthome/core/binding/BindingConfigParseException.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.model.item.binding;
+package org.eclipse.smarthome.core.binding;
 
 public class BindingConfigParseException extends Exception {
