+++ bundles/model/org.eclipse.smarthome.model.item/OSGI-INF/genericitemprovider.xml
@@ -14,6 +14,6 @@
       <provide interface="org.eclipse.smarthome.core.items.ItemProvider"/>
    </service>
    <reference bind="setModelRepository" cardinality="1..1" interface="org.eclipse.smarthome.model.core.ModelRepository" name="ModelRepository" policy="dynamic" unbind="unsetModelRepository"/>
-   <reference bind="addBindingConfigReader" cardinality="0..n" interface="org.eclipse.smarthome.model.item.binding.BindingConfigReader" name="BindingConfigReader" policy="dynamic" unbind="removeBindingConfigReader"/>
+   <reference bind="addBindingConfigReader" cardinality="0..n" interface="org.eclipse.smarthome.core.binding.BindingConfigReader" name="BindingConfigReader" policy="dynamic" unbind="removeBindingConfigReader"/>
    <reference bind="addItemFactory" cardinality="0..n" interface="org.eclipse.smarthome.core.items.ItemFactory" name="ItemFactory" policy="dynamic" unbind="removeItemFactory"/>
 </scr:component>
